+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 #### ADDITIONS
 
-- posts route: fetch post, comment, like, unlike and delete 
+- added README
+- notifications: on like, unlike, comment, user image and delete post
+- user route: added notifications to user when loggen in get authenticated user
+- user route: added two more end points, get user by handle and mark seen notifications read
 
 #### CHANGES
-
-- changes router => route in all routes